% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_density_area.R
\name{poi_density_area}
\alias{poi_density_area}
\title{Compute POI density}
\usage{
poi_density_area(area_sf, poi_list, area_id)
}
\arguments{
\item{area_sf}{object of class sf}

\item{poi_list}{list of POIs}

\item{area_id}{string of column name}
}
\value{
object of class sf
}
\description{
Compute POI density
}
\examples{
poi_list = list("amenity" = "school")
poi_sf <- poi_density_area(districts_lk,poi_list,"ADM2_PCODE")
}
