% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_density_proximity.R
\name{poi_density_proximity}
\alias{poi_density_proximity}
\title{Compute POI density within Specified Proximity}
\usage{
poi_density_proximity(locations_sf, poi_list, location_id, radius = 500)
}
\arguments{
\item{locations_sf}{object of class sf contains locations of interest}

\item{poi_list}{list of POIs}

\item{location_id}{string of location identifier}

\item{radius}{search radius around each location in meters}
}
\value{
object of class sf
}
\description{
Compute POI density within Specified Proximity
}
\examples{
library(ggplot2)
 library(ggspatial)

poi_list = list(
"amenity" = "cafe",
"amenity" = "fast_food")

poi_sf <- poi_density_proximity(com_sec_school,poi_list,"urn",500)
ggplot() +
annotation_map_tile(zoomin = 0, progress = "none")+
geom_sf(data = poi_sf,colour = "red", mapping = aes(fill = poi_density))+
geom_sf(data = com_sec_school)+
theme_void()
}
